Split indexer worker from coordinator
coordinator is pretty bare at this point.
